Remove logging of args for failed iscsi cmds
The iscsiadm functions provide a debug mode that currently logs the
iscsiadm command that was issued and the response.  This is handy for
the lib as it's in beta and we want to make it easy for users to pick up
and start using, BUT it also has a major problem in that it will also
log the arguments passed in for things like CHAP secrets.

This change just removes the debug logging of the iscsiadm cmd args, in
the future if we need/want some debug tools we can implement a more
robust and secure method that's safe and doesn't leak secrets to log
files.

closes #8